Refactor create and update-framework commands.
